+++ utils.go
@@ -103,3 +103,16 @@ func endsWithAny(str string, postfixes []string, cs caseSensitivity) bool {
 	}
 	return false
 }
+
+// file
+
+// Returns if file/dir in path exists.
+func fileExists(path string) bool {
+	if len(path) == 0 {
+		return false
+	}
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		return false
+	}
+	return true
+}